+++ CHANGELOG.md
@@ -1,12 +1,12 @@
 # Changelog
 
 ## 1.0.3
-* Update: Updated for MediaWiki >= 1.38 (z929669)
+* Update: Updated for MediaWiki >= 1.38 compatibility
 * Still works for MediaWiki >= 1.35
 
 ## 1.0.2
-* Update: Updated for IPB 4.x (karu)
-* Update: Changed validations, updated for MW 1.35 (Jordan Songer)
+* Update: Updated for IPB 4.x
+* Update: Changed validations, updated for MW 1.35
 
 ## 1.0.1
 * Bugfix: Users could not edit their signature